Add data type as flow

+++ first_tensorflow.py
@@ -7,8 +7,8 @@ from tensorflow.keras.layers import Dense
 model = Sequential([Dense(units=1, input_shape=[1])])
 model.compile(optimizer='sgd', loss='mean_squared_error')
 
-xs = np.array([-1.0, 0.0, 1.0, 2.0, 3.0, 4.0])
-ys = np.array([-3.0, -1.0, 1.0, 3.0, 5.0, 7.0])
+xs = np.array([-1.0, 0.0, 1.0, 2.0, 3.0, 4.0], dtype=float)
+ys = np.array([-3.0, -1.0, 1.0, 3.0, 5.0, 7.0], dtype=float)
 
 model.fit(xs, ys, epochs=500)
